+++ tests/src/main/java/ee/jakarta/tck/mvc/tests/viewengine/algorithm/ViewEngineAlgorithmController.java
@@ -32,7 +32,7 @@ public class ViewEngineAlgorithmController {
     @GET
     @Path("overwrite-builtin")
     public String overwriteBuiltin() {
-        return "foobar.xhtml";
+        return "foobar.jsp";
     }
 
     @GET
